refactor: move date-fns util into @repo/shared (#1466)
