+++ app/components/Home/components/Section/components/SectionWhitePapers/sectionWhitePapers.tsx
@@ -1,5 +1,4 @@
 import { SectionTitle } from "@/components/Home/components/Section/section.styles";
-import { formatDate } from "@/utils/date-fns";
 import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/typography";
 import {
   CardActionArea,
@@ -7,6 +6,7 @@ import {
   CardMedia,
   Typography,
 } from "@mui/material";
+import { formatDate } from "@repo/shared/utils/date-fns/utils";
 import Link from "next/link";
 import { JSX } from "react";
 import { WHITE_PAPERS } from "./constants";
